directly enter into logs feature added

+++ controllers/deficit_calc.go
@@ -5,12 +5,12 @@ import (
 	"math"
 )
 
-func DeficitCalc(current_weight int, desired_weight int, time_frame int) sql.NullInt32 {
+func DeficitCalc(current_weight int, desired_weight int, time_frame int) sql.NullFloat64 {
 	var difference int = int(math.Floor(math.Abs(float64(current_weight) - float64(desired_weight))))
 	diff_in_kcal := difference * 7716
 	deficit := (diff_in_kcal / time_frame) / 7
-	return sql.NullInt32{
-		Int32: int32(deficit),
+	return sql.NullFloat64{
+		Float64: float64(deficit),
 		Valid: true,
 	}
 }
