+++ sql/schemas/003_daily_nutrition_intake.sql
@@ -4,11 +4,11 @@ CREATE TABLE daily_nutrition_intake(
   id UUID PRIMARY KEY,
   created_at TIMESTAMP NOT NULL,
   user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-  calories INT NOT NULL,
-  carbohydrates INT NOT NULL,
-  protien INT NOT NULL,
-  fat INT NOT NULL,
-  fiber INT NOT NULL
+  calories FLOAT NOT NULL,
+  carbohydrates FLOAT NOT NULL,
+  protien FLOAT NOT NULL,
+  fat FLOAT NOT NULL,
+  fiber FLOAT NOT NULL
 );
 
 -- +goose Down
