+++ sql/schemas/004_total_kcal_intake.sql
@@ -3,9 +3,9 @@
 CREATE TABLE total_calorie_intake(
   id UUID PRIMARY KEY,
   created_at TIMESTAMP NOT NULL,
-  calories INT NOT NULL,
-  total_deficit INT NOT NULL,
-  total_surplus INT NOT NULL,
+  calories FLOAT NOT NULL,
+  total_deficit FLOAT NOT NULL,
+  total_surplus FLOAT NOT NULL,
   user_id UUID NOT NULL REFERENCES users(id) ON DELETE CASCADE
 );
